requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.4
django-oauth-toolkit==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
httpie==3.1.0
idna==3.3
jwcrypto==1.0
multidict==6.0.2
oauthlib==3.2.0
pycparser==2.21
Pygments==2.12.0
PyJWT==1.7.1
PySocks==1.7.1
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.9.1
sqlparse==0.4.2
urllib3==1.26.9
wrapt==1.14.0
